refactor, deals with folders/sections and some odd cases now. checks if file exist/filesize identical

+++ scraper.py
@@ -1,11 +1,16 @@
-from selenium import webdriver
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+
+# Copyright: David Adrian
+# david.adrian@tum.de
+
 import errno
 import os
 import requests
 import argparse
 import getpass
-import menu
-import logging
+
+from selenium import webdriver
 from progressbar import ProgressBar
 # from selenium.webdriver.support.ui import Select # for <SELECT> HTML form
 
@@ -48,6 +53,7 @@ def get_session(driver):
     cookies = driver.get_cookies()
     for cookie in cookies:
         session.cookies.set(cookie['name'], cookie['value'])
+    return session
 
 
 def get_driver(driver_type):
@@ -60,19 +66,6 @@ def get_driver(driver_type):
         driver = webdriver.Firefox()
 
     return driver
-
-
-# But actually only looking at non-hidden is a pain, so i just filter the list
-# def set_semester(semester):
-#     if semester == "all":
-#         # Switch on to view all courses
-#         driver.find_element_by_xpath("//select[@id='coc-filterterm']/option[@value='all']").click()
-#         return True
-#     else:
-#         try:
-#             el = driver.find_element_by_xpath("//select[@id='coc-filterterm']/option[@value=\'" + semester + "\']").click()
-#         except:
-#             print "ERROR: Semester", semester, "does not exist!"
 
 
 def get_courses(driver, semester):
@@ -92,33 +85,31 @@ def get_courses(driver, semester):
     return course_list
 
 
-def download_file(session, path, url):
-    # NOTE the stream=True parameter
-    r = session.get(url, stream=True)
-    total_size = int(r.headers['Content-Length'])
-    fn = os.path.join(path, r.url.split('/')[-1])
-    with open(fn, 'wb') as f:
-        pbar = ProgressBar(maxval=total_size).start()
-        for idx, chunk in enumerate(r.iter_content(chunk_size=1024)):
-            if chunk:  # filter out keep-alive new chunks
-                f.write(chunk)
-                pbar.update(idx * 1024)
-        pbar.finish()
-
-
-def download_files(session, fn_base, files):
+def download_files(session, fn_base, files, force):
     # Download Files
     for file in files:
-        print "Processing: ", file[1], file[2]
-
         # Create Folder if it doesn't exist yet
         path = os.path.join(fn_base, file[0], file[1])
         mkdir_p(path)
 
-        # TODO CHECK IF FILE EXISTS ON FS BEFORE DOWNLOADING
-
         # Download File
-        download_file(session, path, file[3])
+        # NOTE the stream=True parameter
+        # NOTE add the "&redirect=1", to deal with some odd balls
+        r = session.get(file[2] + "&redirect=1", stream=True)
+        fs = int(r.headers['Content-Length'])
+        name = r.headers["Content-Disposition"].split('\"')[-2].decode('utf-8')
+        fn = os.path.join(path, name)
+        if not force and os.path.isfile(fn) and os.stat(fn).st_size == fs:
+            print "   :: File already exists. Skipping", file[1], name, "!"
+        else:
+            print "   :: Downloading:", file[1], name
+            with open(fn, 'wb') as f:
+                pbar = ProgressBar(maxval=fs).start()
+                for idx, chunk in enumerate(r.iter_content(chunk_size=1024)):
+                    if chunk:  # filter out keep-alive new chunks
+                        f.write(chunk)
+                        pbar.update(idx * 1024)
+                pbar.finish()
 
 
 if __name__ == '__main__':
@@ -130,81 +121,89 @@ if __name__ == '__main__':
                         help='Shibboleth User',
                         required=True)
     parser.add_argument('--pw', '-p', type=str, default='',
-                        help='Dev param, password will be visible',
+                        help='(OPTIONAL DEV) Shibboleth Password',
                         required=False)
     parser.add_argument('--semester', '-s', type=str, default='all',
+                        help='Filter courses by semester',
                         required=False)
     parser.add_argument('--driver', '-d', type=str, default='phantomjs',
+                        help='Phantomjs, Chrome, Firefox supported',
+                        required=False)
+    parser.add_argument('--force', '-f', dest='force', action='store_true',
+                        help='Force re-downloads every file',
                         required=False)
     args = parser.parse_args()
 
 
 # Obtain Driver
 print " :: Obtaining Driver"
-driver = get_driver(args.driver)
+driver = get_driver(args.driver.lower())
 
 # We are now on the course pages
-print " :: Login"
+print " :: Login User:", args.user
 tum_login(driver, args.user, args.pw)
 
 # print " :: Setting Semester to ", args.semester
 # set_semester(args.semester)
 
 # Get all courses and filter by semester/all
-print " :: Getting Courses of", args.semester
+print "\n :: Getting Courses of", args.semester
 courses = get_courses(driver, args.semester)
-# print courses
+print "   :: Found:", len(courses), "courses!"
 
-print " :: Collecting File Information"
-
-# now process all the courses
-
+print "\n :: Collecting File Information"
+# First get all special cases pre-processed
+## TODO: Should be recursive just to make sure!
 folder_queue = []
 section_queue = []
-
 for course in courses:
-    print course[0], course[1]
+    print "   ::", course[0], course[1]
     driver.get(course[2])
 
+    # 2. Get folders
     try:
-        # 2. Recurse through folder
         # TODO: Timeout problems?
         folders = driver.find_elements_by_class_name("modtype_folder")
         for folder in folders:
             folder_href = folder.find_element_by_xpath(".//a").get_attribute('href')
-            folder_name = make_fs_string(file.find_element_by_class_name("instancename").text)
+            folder_name = make_fs_string(folder.find_element_by_class_name("instancename").text)
             folder_queue.append(((course[0], os.path.join(course[1], folder_name), folder_href)))
     except:
-        print " :: :: Folder Search Faield!!!"
+        print "   :: Folder Search Failed!!!"
 
     # 3. Deal with sections (similar to folder)
-    # try:
-    #     # 2. Recurse through folder
-    #     # TODO: Timeout problems?
-    #     folders = driver.find_elements_by_class_name("section-summary")
-    #     for folder in folders:
-    #         link_el = folder.find_element_by_xpath(".//a")
-    #         folder_href = link_el.get_attribute('href')
-    #         folder_name = make_fs_string(link_el.text)
-    #         section_queue.append(((course[0], os.path.join(course[1], folder_name), folder_href)))
-    # except:
-    #     print " :: :: Folder Search Faield!!!"
+    try:
+        # 2. Recurse through folder
+        # TODO: Timeout problems?
+        folders = driver.find_elements_by_class_name("section-summary")
+        for folder in folders:
+            link_el = folder.find_element_by_xpath(".//a")
+            folder_href = link_el.get_attribute('href')
+            folder_name = make_fs_string(link_el.text)
+            section_queue.append(((course[0], os.path.join(course[1], folder_name), folder_href)))
+    except:
+        print "   :: Folder Search Failed!!!"
 
-
-file_store = []
+# Now treat everything as page and look for files
+parsed_files = []
 pages = courses + folder_queue + section_queue
+# pages = folder_queue
 for page in pages:
     # 1. Find single files
-    driver.get(course[2])
+    driver.get(page[2])
     try:
-        files = driver.find_elements_by_class_name("modtype_resource")
+        files = driver.find_elements_by_class_name("modtype_resource") + driver.find_elements_by_class_name("fp-filename-icon")
         for file in files:
-            file_href = file.find_element_by_xpath(".//a").get_attribute('href')
-            file_name = make_fs_string(file.find_element_by_class_name("instancename").text)
-            file_store.append((course[0], course[1], file_name, file_href))
+            try:
+                file_href = file.find_element_by_xpath(".//a").get_attribute('href')
+                parsed_files.append((page[0], page[1], file_href))
+            except:
+                # wasnt a file, only icon
+                pass
     except:
-        print " :: :: No Files Found!"
+        print "   :: No Files Found!"
 
-
-
-print file_store
+# Download files if they don't exist.
+print "\n :: Collected", len(parsed_files), "files"
+print " :: Downloading files (please wait)"
+download_files(get_session(driver), args.fn_out, parsed_files, args.force)
